Fixed a bug in the graph rendering and made metricType synced one-way from the sorter
There was a bug when if you rendered a graph and then changed the sorting type, the hedge fund’s info would be repositioned, but the graph would stay stuck in place.

When it comes to sorting, now if you change the sorting type, the sorting type changes for all cards to improve the user experience. However, this does not go the other way around as that would make it very annoying to use the app, since the card you would be changing the metric for would get lost in the abyss.