Add filtering options, update todos.
This was urgent since there are too many funds at this point and over ~200, the site gets very laggy and takes up gigabytes of ram. To make it runnable on my phone this had to be done. Also, the filter by months allows to filter out the new funds with non-established performance metrics.